requirements.txt
asgiref==3.7.2
black==24.4.2
bleach==6.1.0
click==8.1.7
Django==5.0.2
django-environ==0.11.2
django-pandas==0.6.6
django-summernote==0.8.20.0
django-widget-tweaks==1.5.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.2
plotly==5.18.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
tzdata==2024.1
webencodings==0.5.1
